Implement the `dotenv!` macro
The meat of this macro is absurdly simple, most of the code here is just
machinery for use on both stable and nighlty.

The macro is identical to the `env!` macro, but it calls
`dotenv::dotenv()` for you at compile time. If a `.env` file is found,
but fails to parse, the code will not compile. If a `.env` file is not
found, the behavior is simply that of `env!`, as `.env` files are meant
to be optional.

`dotenv_codegen` and `dotenv_macros` are separate crates, because we
need the format to be `rlib` on stable, but `dylib` on nightly
(specifying both can cause some weird issues when building downstream).

This doesn't include tests for the macro, as it would require jumping
through some hoops to get a simple test working, but I've tested this on
Diesel and it works great. A test can easily be added later.