Add PathBuf return to from_filename() and dotenv()
This implements #22, with the exception of returning a PathBuf instead of a String, and not adding the return to from_path.

PathBuf should not contain less information, and will be practically free to add because we already create a new PathBuf when joining files together. Returning a String on the other hand would require allocating additional memory.